TestTools/python/requirements.txt
crc8
pyserial==3.4
